requirements.txt
PyGithub==2.2.0
requests==2.32.0
python-dotenv==1.0.1
aiohttp==3.10.0
Flask==3.0.0
flask-cors==4.0.0
pytest==7.4.0
pytest-cov==4.1.0

# RAG dependencies
langchain==0.1.0
chromadb==0.4.22

# FastAPI + WebSocket
fastapi==0.109.0
uvicorn==0.27.0
python-socketio==5.11.0
httpx==0.26.0
